Add WeatherStore

+++ data/weather/src/commonMain/kotlin/io/bitpogo/keather/data/weather/model/store/SaveableForecast.kt
@@ -13,5 +13,5 @@ internal data class SaveableForecast(
     val averageTemperatureInCelsius: Double,
     val maximumWindSpeedInKilometerPerHour: Double,
     val precipitationInMillimeter: Double,
-    val rainPossibility: Int,
+    val rainPossibility: Long,
 )
